Add initial setup operation
This operation checks that the nodes are in a healthy condition wrt to number of processes of the current user. There is a bug or misbehaviour in Xcode/Simulators that can freeze the entire machine after running many test suites (weeks of uptime) because the simulator takes lots of processes.